工作流库0216/T01-每日数据获取任务/requirements.txt
# T01-每日数据获取任务 依赖
# 任务ID: T01
# 创建时间: 2026-02-14

# ============================================
# 核心依赖
# ============================================
pandas>=1.5.0
numpy>=1.21.0

# ============================================
# 数据库
# ============================================
pymysql>=1.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# ============================================
# THS API (同花顺iFinD)
# ============================================
# 注意: iFinDPy需要单独安装，请联系同花顺获取安装包
# iFinDPy

# ============================================
# 工具库
# ============================================
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# ============================================
# 可选依赖
# ============================================
# matplotlib>=3.5.0  # 可视化
# openpyxl>=3.0.0    # Excel处理
